refactor: simplify api functions and files
